Write new script for extracting keycodes
The new script runs on Python 3. The output uses the standard CSV
format, rather than using tabs, so it should be more interoperable with
other systems. A header with column names is added to the output. Linux
keycodes are taken from input-event-codes.h instead of trying to read
XKB configuration files.